requirements.txt
asgiref==3.10.0
Django==5.2.7
django-jazzmin==3.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
